requirements_colab.txt
numpy
torch
torchvision
pandas
pillow
tqdm
